+++ generator/lib/src/meassure.dart
@@ -1,0 +1,59 @@
+////////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////
+import 'dart:async';
+
+import 'package:build/build.dart'; // for shared logging instance log
+
+String humanReadable(Duration duration) {
+  // Added microseconds
+  if (duration < const Duration(milliseconds: 1)) {
+    return '${duration.inMicroseconds}μs';
+  }
+  if (duration < const Duration(seconds: 1)) {
+    return '${duration.inMilliseconds}ms';
+  }
+  if (duration < const Duration(minutes: 1)) {
+    return '${(duration.inMilliseconds / 1000.0).toStringAsFixed(1)}s';
+  }
+  if (duration < const Duration(hours: 1)) {
+    final minutes = duration.inMinutes;
+    final remaining = duration - Duration(minutes: minutes);
+    return '${minutes}m ${remaining.inSeconds}s';
+  }
+  final hours = duration.inHours;
+  final remaining = duration - Duration(hours: hours);
+  return '${hours}h ${remaining.inMinutes}m';
+}
+
+// NOTE: This is copied from `package:build_runner_core`.
+FutureOr<T> meassure<T>(FutureOr<T> Function() action,
+    {String tag = '', repetitions = 1}) async {
+  return [
+    for (int i = 0; i < repetitions; ++i)
+      await (() async {
+        final stopwatch = Stopwatch()..start();
+        try {
+          return await action();
+        } finally {
+          stopwatch.stop();
+          final time = humanReadable(stopwatch.elapsed);
+          log.info('[$tag ($i)] completed, took $time');
+        }
+      })()
+  ].last;
+}
